Remove unused settings.h & make window resizing work

+++ include/elements.h
@@ -1,6 +1,5 @@
 #ifndef ELEMENTS_H
 # define ELEMENTS_H
-# include "settings.h"
 # include "ftvec3.h"
 
 //The rgb: [0-255]
@@ -18,7 +17,7 @@ typedef t_vec3 t_normal;
 
 //The lighting ration: [0.0, 1.0]
 typedef struct s_ambient_light {
-	FLOAT	lighting_ratio;
+	double	lighting_ratio;
 	t_vec3	color;
 } t_ambient_light;
 
@@ -33,13 +32,13 @@ typedef struct s_camera_o {
 //RBG not used in mandatory part
 typedef struct	s_light {
 	t_point		loc;
-	FLOAT		brightness;
+	double		brightness;
 	t_vec3		color;
 } t_light;
 
 typedef struct	s_sphere {
 	t_point	loc;
-	FLOAT	diameter;
+	double	diameter;
 	t_vec3	color;
 } t_sphere;
 
@@ -52,8 +51,8 @@ typedef struct	s_plane {
 typedef struct	s_cylinder{
 	t_point		loc;
 	t_normal	normal;
-	FLOAT		diameter;
-	FLOAT		height;
+	double		diameter;
+	double		height;
 	t_vec3		color;
 } t_cylinder;
